pyproject.toml
[build-system]
requires = ["setuptools>=77.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "frontend-design-loop-mcp"
version = "1.0.0"
description = "Frontend Design Loop MCP - agent-first patch evaluation, preview, and screenshot tooling"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Alexander Burkhart"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["mcp", "claude-code", "agentic", "automation", "code-review", "patch", "vision"]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "mcp>=1.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "google-auth>=2.23.0",
    "playwright>=1.40.0",
    "aiosqlite>=0.19.0",
    "jsonschema>=4.20.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "google-cloud-storage>=2.13.0",
]

[project.urls]
Homepage = "https://github.com/alexalexalex222/frontend-design-loop-mcp"
Repository = "https://github.com/alexalexalex222/frontend-design-loop-mcp"
Issues = "https://github.com/alexalexalex222/frontend-design-loop-mcp/issues"
Documentation = "https://github.com/alexalexalex222/frontend-design-loop-mcp/blob/main/docs/FRONTEND_DESIGN_LOOP_MCP.md"
Changelog = "https://github.com/alexalexalex222/frontend-design-loop-mcp/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "build>=1.2.0",
    "twine>=5.0.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "types-pyyaml",
    "types-jsonschema",
]

[project.scripts]
frontend-design-loop-mcp = "frontend_design_loop_mcp.mcp_server:main"
frontend-design-loop-setup = "frontend_design_loop_mcp.setup:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = [
    "frontend_design_loop_mcp",
    "frontend_design_loop_mcp.*",
    "frontend_design_loop_core",
    "frontend_design_loop_core.*",
]

[tool.setuptools.package-data]
frontend_design_loop_mcp = [
    "assets/config/*.yaml",
    "assets/prompts/*",
    "assets/templates/nextjs_app_router_tailwind/*",
    "assets/templates/nextjs_app_router_tailwind/app/*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
